+++ core/testing/build.gradle.kts
@@ -9,7 +9,6 @@ kotlin {
     sourceSets {
         commonMain.dependencies {
             implementation(libs.kotlin.test)
-            implementation(libs.kotlinx.coroutines.test)
             implementation(libs.assertk)
             implementation(libs.turbine)
         }
