+++ core/ui/contract/src/commonTest/kotlin/net/thunderbird/core/ui/contract/mvi/BaseViewModelTest.kt
@@ -11,7 +11,7 @@ import kotlin.test.Test
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
-import net.thunderbird.core.testing.coroutines.MainDispatcherHelper
+import net.thunderbird.core.ui.testing.coroutines.MainDispatcherHelper
 
 class BaseViewModelTest {
