+++ core/ui/testing/src/commonMain/kotlin/net/thunderbird/core/ui/testing/coroutines/MainDispatcherHelper.kt
@@ -1,4 +1,4 @@
-package net.thunderbird.core.testing.coroutines
+package net.thunderbird.core.ui.testing.coroutines
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -31,16 +31,16 @@ import kotlinx.coroutines.test.setMain
  * Use [kotlinx.coroutines.test.UnconfinedTestDispatcher] if you want to execute coroutines immediately without any
  * scheduling, which is useful for simple tests. However, if you need more control over the execution of coroutines
  * (e.g., to test delays, timeouts, or to ensure that certain code runs in a specific order), stay with the default
- *  [kotlinx.coroutines.test.StandardTestDispatcher].
+ *  [StandardTestDispatcher].
  *
  * @param testDispatcher The dispatcher to set as the main dispatcher. Defaults to [StandardTestDispatcher].
  */
-class MainDispatcherHelper(
-    val testDispatcher: TestDispatcher = StandardTestDispatcher(),
+public class MainDispatcherHelper(
+    public val testDispatcher: TestDispatcher = StandardTestDispatcher(),
 ) {
     @OptIn(ExperimentalCoroutinesApi::class)
-    fun setUp() = Dispatchers.setMain(testDispatcher)
+    public fun setUp(): Unit = Dispatchers.setMain(testDispatcher)
 
     @OptIn(ExperimentalCoroutinesApi::class)
-    fun tearDown() = Dispatchers.resetMain()
+    public fun tearDown(): Unit = Dispatchers.resetMain()
 }
