+++ feature/account/server/validation/build.gradle.kts
@@ -27,6 +27,7 @@ dependencies {
     implementation(projects.feature.account.oauth)
     implementation(projects.feature.account.server.certificate)
 
+    testImplementation(projects.core.ui.testing)
     testImplementation(projects.core.ui.compose.testing)
 }
