+++ feature/account/settings/impl/build.gradle.kts
@@ -43,6 +43,7 @@ dependencies {
     debugImplementation(projects.core.ui.setting.implDialog)
 
     testImplementation(projects.core.logging.testing)
+    testImplementation(projects.core.ui.testing)
     testImplementation(projects.core.ui.compose.testing)
 }
