+++ feature/account/setup/build.gradle.kts
@@ -41,6 +41,7 @@ dependencies {
 
     implementation(projects.feature.thundermail.api)
 
+    testImplementation(projects.core.ui.testing)
     testImplementation(projects.core.logging.testing)
     testImplementation(projects.core.ui.compose.testing)
