+++ feature/funding/googleplay/src/test/kotlin/net/thunderbird/feature/funding/googleplay/ui/reminder/ActivityLifecycleObserverTest.kt
@@ -13,7 +13,7 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
 import net.thunderbird.core.testing.TestClock
-import net.thunderbird.core.testing.coroutines.MainDispatcherHelper
+import net.thunderbird.core.ui.testing.coroutines.MainDispatcherHelper
 
 @OptIn(ExperimentalTime::class)
 class ActivityLifecycleObserverTest {
