+++ feature/navigation/drawer/dropdown/build.gradle.kts
@@ -32,6 +32,7 @@ dependencies {
     implementation(projects.legacy.ui.folder)
     implementation(projects.core.featureflag)
 
+    testImplementation(projects.core.ui.testing)
     testImplementation(projects.core.ui.compose.testing)
     testImplementation(projects.core.logging.testing)
