+++ legacy/core/build.gradle.kts
@@ -51,6 +51,7 @@ dependencies {
     implementation(projects.feature.mail.message.reader.api)
 
     testImplementation(projects.core.testing)
+    testImplementation(projects.core.ui.testing)
     testImplementation(projects.core.android.testing)
     testImplementation(projects.core.logging.testing)
     testImplementation(projects.feature.telemetry.noop)
