+++ legacy/core/src/test/java/com/fsck/k9/notification/SingleMessageNotificationCreatorTest.kt
@@ -20,7 +20,7 @@ import net.thunderbird.core.android.testing.MockHelper.mockBuilder
 import net.thunderbird.core.android.testing.RobolectricTest
 import net.thunderbird.core.preference.notification.NotificationPreference
 import net.thunderbird.core.preference.notification.NotificationPreferenceManager
-import net.thunderbird.core.testing.coroutines.MainDispatcherHelper
+import net.thunderbird.core.ui.testing.coroutines.MainDispatcherHelper
 import org.junit.After
 import org.junit.Before
 import org.junit.Test
